requirements.txt
altgraph==0.17.5
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==4.2
django-celery-beat==2.8.1
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
macholib==1.16.4
packaging==25.0
paho-mqtt==2.1.0
postgres==4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
psycopg2-pool==1.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
